lab1/client/requirements.txt
python-socketio==5.4.1
asyncio==3.4.3
aiohttp==3.7.4.post0